S-14a extract status and costs pages to web/modules/status-costs.js
Move loadStatus (Statuspage/incident viewer) and loadCosts (cost ledger
summary) from app.js into status-costs.js. Both pages are standalone with no
cross-page dependencies beyond t() and a local escapeHtml. initStatusCosts()
wires the two refresh-button event listeners.
